Update processing output and optimize tests
Reduced the number of random rows for demo data from 10 to 2 and updated the timer output in Program.cs for more accuracy. This means the elapsed time is now shown in total microseconds and milliseconds. Commented out the table generation code to focus on raw data processing benchmarks. Added ARM64 target option for Release configuration in CsvKitDemo.csproj. Renamed test methods for clarity in ComponentTests.cs.